move Slicer and Worker context to their files
