+++ packages/teraslice-test-harness/src/index.ts
@@ -1,3 +1,5 @@
-import JobHarness from './job-harness';
+import SlicerTestHarness from './slicer-test-harness';
+import WorkerTestHarness from './worker-test-harness';
 
-export { JobHarness };
+export * from './interfaces';
+export { SlicerTestHarness, WorkerTestHarness };
