+++ packages/teraslice-test-harness/test/index-spec.ts
@@ -2,7 +2,9 @@ import 'jest-extended';
 import * as index from '../src';
 
 describe('index', () => {
-    it('should have JobHarness', () => {
-        expect(index).toHaveProperty('JobHarness');
+    it('should have the test harnesses', () => {
+        expect(index).not.toHaveProperty('JobTestHarness');
+        expect(index).toHaveProperty('SlicerTestHarness');
+        expect(index).toHaveProperty('WorkerTestHarness');
     });
 });
